1. update post implemented 2. user/guest mode implemented

+++ XamarinBlogEducation.DataAccess/Repositories/Interfaces/IPostsRepository.cs
@@ -11,7 +11,7 @@ namespace XamarinBlogEducation.DataAccess.Repositories.Interfaces
         Task<IEnumerable<Post>> GetByKey(string key);
         Task<IEnumerable<Post>> GetByDate(DateTime CreationDate);
         Task<IEnumerable<Post>> GetByCategory(long categoryId);
-        Task<Post> GetPost(int id);
+        Task<Post> GetPost(long id);
         Task<IEnumerable<Category>> GetAllCategories();
         Task<IEnumerable<Post>> GetByAuthor(string userId);
     }
